IRC petal working somewhat. A testing interface is included.
Start bin/licd and you'll get a prompt. <lic> specifies that you're
talking to the lic item. Type `/to irc/freenode/#oftn` to switch to #oftn.
Then, `/join` will join the channel. If you type a message, it will be
sent as a "message" command to the current item.

You can also play with the config API a bit. `/to
lic/config/irc/defaults/nick` and then `/set "devyn|lic"`, for example.

When done, either just press control-C, or type `/to lic` then `/quit`.