Fix: don't call Server#shutdown multiple times in Hub#shutdown
Hub#shutdown would first shut down all the petals, and then check the amount of open petals. This works fine when the petals aren't all gone when exit is called, but when there's just config, the num becomes 0 before that check, so exit() is called multiple times, resulting in strange behavior.